Needs to extend also the init_containers not just override (e.g. git sync and similar init_containers will stay)
With that change if you define a pod_override and specifies the "init_containers" attribute it will extend the base init_containers with the new ones instead of overwriting them. That is to keep the dag git sync init_containers or other pre-defined init_containers based on the template or the configuration.